+++ app/Controller/Setting/TimedTaskController.php
@@ -8,6 +8,8 @@ use App\Constants\StatusCode;
 use App\Controller\AbstractController;
 use App\Foundation\Annotation\Explanation;
 use App\Foundation\Utils\Cron;
+use App\Model\Auth\User;
+use App\Model\Laboratory\FriendRelation;
 use App\Model\Setting\TimedTask;
 use Hyperf\Di\Annotation\Inject;
 use Hyperf\HttpServer\Annotation\Controller;
@@ -222,8 +224,14 @@ class TimedTaskController extends AbstractController
      */
     public function destroy(int $id)
     {
-        if (!intval($id)) $this->throwExp(StatusCode::ERR_VALIDATION, '参数错误');
-        if (!TimedTask::destroy($id)) $this->throwExp(StatusCode::ERR_EXCEPTION, '删除失败');
+        if ($id == 0) {
+            $idArr = $this->request->input('id') ?? [];
+            if (empty($idArr) || !is_array($idArr)) $this->throwExp(StatusCode::ERR_VALIDATION, '参数类型不正确');
+            if (!TimedTask::whereIn('id', $idArr)->delete()) $this->throwExp(StatusCode::ERR_EXCEPTION, '删除失败');
+        }else {
+            if (!intval($id)) $this->throwExp(StatusCode::ERR_VALIDATION, '参数错误');
+            if (!TimedTask::destroy($id)) $this->throwExp(StatusCode::ERR_EXCEPTION, '删除失败');
+        }
 
         return $this->successByMessage('删除定时任务成功');
     }
